distsql: merge separate util test files in distsqlrun.
Release notes: none
